+++ tests/c/module.py
@@ -1,0 +1,3 @@
+class Module:
+    def __init__(self, a: int) -> None:
+        self.a = a