+++ yatiml/constructors.py
@@ -10,7 +10,7 @@ from typing_extensions import TYPE_CHECKING, Type
 import yaml
 
 from yatiml.exceptions import RecognitionError
-from yatiml.introspection import class_subobjects
+from yatiml.introspection import class_subobjects, init_function
 from yatiml.util import (
         bool_union_fix, diagnose_extraneous_key, diagnose_missing_key,
         generic_type_args, is_generic_sequence, is_generic_mapping,
@@ -25,13 +25,15 @@ logger = logging.getLogger(__name__)
 class Constructor:
     """A constructor for user classes to register with YAML."""
 
-    def __init__(self, class_: Type) -> None:
+    def __init__(self, class_: Type, reg_class_names: Dict[Type, str]) -> None:
         """Create a constructor
 
         Args:
             class_: The class that this is a constructor for.
+            reg_class_names: Table of names to show for registered classes.
         """
         self.class_ = class_
+        self._reg_class_names = reg_class_names
 
     def __call__(self, loader: 'Loader',
                  node: yaml.Node) -> Generator[Any, None, None]:
@@ -71,9 +73,11 @@ class Constructor:
 
         self.__loader = loader
 
+        initialiser = init_function(self.class_)
+
         # figure out which keys are extra and strip them of tags
         # to prevent constructing objects we haven't type checked
-        argspec = inspect.getfullargspec(self.class_.__init__)
+        argspec = inspect.getfullargspec(initialiser)
         self.__strip_extra_attributes(node, argspec.args)
 
         # create object and let yaml lib construct subobjects
@@ -94,15 +98,18 @@ class Constructor:
             if '_yatiml_extra' in argspec.args:
                 attrs = self.__split_off_extra_attributes(
                     mapping, argspec.args)
-                new_obj.__init__(**attrs)
+                initialiser(new_obj, **attrs)
 
             else:
-                new_obj.__init__(**mapping)
+                initialiser(new_obj, **mapping)
 
         except Exception as e:
             raise RecognitionError(
                     'An error occurred:\n{}\n{}'.format(node.start_mark, e))
         logger.debug('Done constructing {}'.format(self.class_.__name__))
+
+    def __ttd(self, t: Type) -> str:
+        return type_to_desc(t, self._reg_class_names)
 
     def __split_off_extra_attributes(self, mapping: Dict,
                                      known_attrs: List[str]) -> Dict:
@@ -198,8 +205,8 @@ class Constructor:
                 raise RecognitionError(
                         '{}\nAttribute "{}" is {}, expected {}'.format(
                             node.start_mark, name,
-                            type_to_desc(type(mapping[name])),
-                            type_to_desc(type_)))
+                            self.__ttd(type(mapping[name])),
+                            self.__ttd(type_)))
 
     def __type_check_attributes(self, node: yaml.Node, mapping: Dict,
                                 argspec: inspect.FullArgSpec) -> None:
@@ -239,8 +246,8 @@ class Constructor:
                             '{}\nExpected attribute "{}" to be {} but it is {}'
                             ).format(
                                     value_node.start_mark, key,
-                                    type_to_desc(argspec.annotations[key]),
-                                    type_to_desc(type(value))))
+                                    self.__ttd(argspec.annotations[key]),
+                                    self.__ttd(type(value))))
 
     def __strip_extra_attributes(self, node: yaml.Node,
                                  known_attrs: List[str]) -> None:
@@ -257,11 +264,13 @@ class Constructor:
         """
         known_keys = list(known_attrs)
         if 'self' not in known_keys:
-            raise RuntimeError('The __init__ method of {} does not have a'
-                               ' "self" attribute! Please add one, this is'
-                               ' not a valid constructor.'.format(
-                                   self.class_.__name__))
+            raise RuntimeError(
+                    'The __init__ (or _yatiml_init, if present) method of {}'
+                    ' does not have a "self" attribute! Please add one, this'
+                    ' is not a valid constructor.'.format(
+                        self.class_.__name__))
         known_keys.remove('self')
+
         if '_yatiml_extra' in known_keys:
             known_keys.remove('_yatiml_extra')
 
@@ -281,13 +290,15 @@ class EnumConstructor:
     i.e. classes derived from enum.Enum.
     """
 
-    def __init__(self, class_: Type) -> None:
+    def __init__(self, class_: Type, reg_class_names: Dict[Type, str]) -> None:
         """Create a constructor
 
         Args:
             class_: The class that this is a constructor for.
+            reg_class_names: Table of names to show for registered classes.
         """
         self.class_ = class_
+        self._reg_class_names = reg_class_names
 
     def __call__(self, loader: 'Loader',
                  node: yaml.Node) -> Generator[Any, None, None]:
@@ -312,7 +323,7 @@ class EnumConstructor:
 
         msg = (
                 'An error occurred:\n{}\nExpected a string matching {}.'
-                ).format(node.start_mark, type_to_desc(self.class_))
+                ).format(node.start_mark, self.__ttd(self.class_))
 
         if (
                 not isinstance(node, yaml.ScalarNode) or
@@ -327,6 +338,9 @@ class EnumConstructor:
             raise RecognitionError(msg)
         yield new_obj
 
+    def __ttd(self, t: Type) -> str:
+        return type_to_desc(t, self._reg_class_names)
+
 
 class UserStringConstructor:
     """A constructor for user-defined string classes to register with YAML.
@@ -336,13 +350,15 @@ class UserStringConstructor:
     collections.UserString.
     """
 
-    def __init__(self, class_: Type) -> None:
+    def __init__(self, class_: Type, reg_class_names: Dict[Type, str]) -> None:
         """Create a constructor
 
         Args:
             class_: The class that this is a constructor for.
+            reg_class_names: Table of names to show for registered classes.
         """
         self.class_ = class_
+        self._reg_class_names = reg_class_names
 
     def __call__(self, loader: 'Loader',
                  node: yaml.Node) -> Generator[Any, None, None]:
@@ -369,7 +385,7 @@ class UserStringConstructor:
                 node.value, str):
             raise RecognitionError(
                 ('{}\nExpected a string matching {}.').format(
-                    node.start_mark, type_to_desc(self.class_)))
+                    node.start_mark, self.__ttd(self.class_)))
 
         # PyYAML expects us to yield an incomplete object, but strings are
         # immutable, so we'll have to make the whole thing right away.
@@ -379,6 +395,9 @@ class UserStringConstructor:
             raise RecognitionError(
                     'An error occurred:\n{}\n{}'.format(node.start_mark, e))
         yield new_obj
+
+    def __ttd(self, t: Type) -> str:
+        return type_to_desc(t, self._reg_class_names)
 
 
 class PathConstructor: